Moves FormFieldCheckbox
