+++ src/forms/FormFieldRadio/FormFieldRadioTypes.tsx
@@ -1,5 +1,5 @@
 import { MosaicLabelValue } from "@root/types";
-import { StrictUnion } from "../FormFieldCheckbox";
+import { StrictUnion } from "@root/components/Field/FormFieldCheckbox";
 import { FieldDefBase } from "@root/components/Field";
 
 type FormFieldRadioLocalOptions = {
